Move gcpservice to gcp package
