+++ service/src/main/java/bio/terra/axonserver/service/iam/SamService.java
@@ -1,7 +1,7 @@
 package bio.terra.axonserver.service.iam;
 
 import bio.terra.axonserver.app.configuration.SamConfiguration;
-import bio.terra.axonserver.service.cloud.aws.GcpService;
+import bio.terra.axonserver.service.cloud.gcp.GcpService;
 import bio.terra.common.iam.BearerToken;
 import bio.terra.common.sam.exception.SamExceptionFactory;
 import org.broadinstitute.dsde.workbench.client.sam.ApiClient;
